requirements.txt
matplotlib>=3.8.0
google-cloud-storage>=2.14.0
